requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
dateparser==1.1.7
dateutils==0.6.12
dbutils==3.0.2
django==4.1.7
django-annoying==0.10.6
django-crontab==0.7.1
django-db-connection-pool==1.2.2
django-debug-toolbar==3.8.1
fiona==1.9.1
geopandas==0.12.2
greenlet==2.0.2
idna==3.4
munch==2.5.0
mysqlclient==2.1.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pillow==9.4.0
pymysql==1.0.2
pyproj==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
scipy==1.10.1
setuptools==67.5.1
shapely==2.0.1
six==1.16.0
sqlalchemy==2.0.5.post1
sqlparse==0.4.3
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
